requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
stripe>=7.0
httpx>=0.27
